client_tools/alliance_scraper/requirements.txt
pillow>=10.0
numpy>=1.26
opencv-python>=4.9
imagehash>=4.3
rapidocr-onnxruntime>=1.3
easyocr>=1.7
supabase>=2.0
python-dotenv>=1.0
requests>=2.30
